requirements.txt
rich
psutil
cryptography
fastapi
uvicorn
jinja2
Pillow
Faker
python-multipart

